+++ src/app/(dashboard)/(blog)/posts/(routes)/comments/components/filters/filters.tsx
@@ -1,0 +1,11 @@
+import ListFilters from '@/components/list/filters/filters';
+
+import PostsCommentsListFilterStatus from './status';
+
+export default function PostsCommentsListFilters() {
+  return (
+    <ListFilters>
+      <PostsCommentsListFilterStatus />
+    </ListFilters>
+  );
+}